Introduce AbstractElement and AbstractElementBuilder to reduce code duplication.
