+++ pumlgenerator/src/main/kotlin/uml/element/ClassElement.kt
@@ -15,74 +15,20 @@ import uml.className
 import uml.fullQualifiedName
 import java.util.Collections.addAll
 
-data class ClassElement(
-    override val elementName: String,
-    override val elementAlias: String,
-    val uniqueIdentifier: String,
-    val attributes: List<Field>,
-    val functions: List<Method>,
+class ClassElement(
+    elementName: String,
+    elementAlias: String,
+    uniqueIdentifier: String,
+    attributes: List<Field>,
+    functions: List<Method>,
     val isSealedClass: Boolean,
     val isDataClass: Boolean,
-    val isShell: Boolean
-) : DiagramElement() {
-    override val comment: String = "'$uniqueIdentifier"
+    isShell: Boolean
+) : AbstractElement(elementName, elementAlias, uniqueIdentifier, attributes, functions, isShell) {
+
     override val elementKind: ElementKind = ElementKind.CLAZZ(isSealedClass, isData = isDataClass)
-    override fun getContent(indent: String): String {
-        val shellString = if (isShell) DiagramElement.shellString else ""
-        val attributesString = attributes
-            .takeIf { it.isNotEmpty() }
-            ?.let { it.joinToString(separator = "\n") { "$indent${it.render()}" } }
-            ?: ""
-        val functionsString = functions
-            .takeIf { it.isNotEmpty() }
-            ?.let { it.joinToString(separator = "\n") { "$indent${it.render()}" } }
-            ?: ""
-        return """
-$shellString
-$attributesString
-$functionsString
-        """
-    }
 
-    class Builder(override val clazz: KSClassDeclaration, override var isShell: Boolean, override val options: Options, val logger: KSPLogger? = null) : DiagramElement.Builder<ClassElement> {
-        val companionObject = clazz.declarations.filter { it is KSClassDeclaration && it.isCompanionObject }.map { it as? KSClassDeclaration }.firstOrNull()
-
-        override val extensionProperties: MutableList<KSPropertyDeclaration> = mutableListOf()
-
-        override val allProperties: List<KSPropertyDeclaration>
-            get() = buildList {
-                if (!isShell) {
-                    val validCompanionObjectProperties = companionObject?.getAllProperties()?.filterPropertiesByOptions(clazz, options, logger) ?: emptySequence()
-                    addAll(validCompanionObjectProperties)
-
-                    val validProperties = clazz.getAllProperties().filterPropertiesByOptions(clazz, options, logger)
-                    addAll(validProperties)
-
-                    val otherProperties = clazz.declarations.filterIsInstance<KSPropertyDeclaration>().filterNot { it in clazz.getAllProperties() }.filterPropertiesByOptions(clazz, options, logger)
-                    addAll(otherProperties)
-                }
-
-                addAll(extensionProperties)
-            }.toMutableList()
-
-        override val extensionFunctions: MutableList<KSFunctionDeclaration> = mutableListOf()
-
-        override val allFunctions: List<KSFunctionDeclaration>
-            get() = buildList {
-                if (!isShell) {
-                    val validCompanionObjectFunctions = companionObject?.getAllFunctions()?.filterFunctionsByOptions(clazz, options, logger) ?: emptySequence()
-                    addAll(validCompanionObjectFunctions)
-
-                    val validFunctions = clazz.getAllFunctions().filterFunctionsByOptions(clazz, options, logger)
-                    addAll(validFunctions)
-
-                    val otherFunctions = clazz.declarations.filterIsInstance<KSFunctionDeclaration>().filterNot { it in clazz.getAllFunctions() }.filterFunctionsByOptions(clazz, options, logger)
-                    addAll(otherFunctions)
-                }
-
-                addAll(extensionFunctions)
-            }.toMutableList()
-
+    class Builder(clazz: KSClassDeclaration, isShell: Boolean, options: Options, logger: KSPLogger? = null) : AbstractElementBuilder(clazz, isShell, options, logger) {
 
         override fun build(): ClassElement? {
             return if (options.isValid(clazz, logger)) {
@@ -104,5 +50,6 @@ $functionsString
         override fun toString(): String {
             return "ClassBuilder(clazz=${clazz.fullQualifiedName})"
         }
+
     }
 }
