+++ pumlgenerator/src/main/kotlin/uml/element/EnumElement.kt
@@ -4,10 +4,6 @@ import Options
 import com.google.devtools.ksp.processing.KSPLogger
 import com.google.devtools.ksp.symbol.ClassKind
 import com.google.devtools.ksp.symbol.KSClassDeclaration
-import com.google.devtools.ksp.symbol.KSFunctionDeclaration
-import com.google.devtools.ksp.symbol.KSPropertyDeclaration
-import filterFunctionsByOptions
-import filterPropertiesByOptions
 import isValid
 import uml.DiagramElement
 import uml.ElementKind
@@ -15,16 +11,15 @@ import uml.className
 import uml.fullQualifiedName
 
 
-data class EnumElement(
-    val uniqueIdentifier: String,
-    override val elementName: String,
-    override val elementAlias: String,
-    val attributes: List<Field>,
-    val functions: List<Method>,
+class EnumElement(
+    uniqueIdentifier: String,
+    elementName: String,
+    elementAlias: String,
+    attributes: List<Field>,
+    functions: List<Method>,
     val members: List<String>,
-    val isShell: Boolean
-) : DiagramElement() {
-    override val comment: String = "'$uniqueIdentifier"
+    isShell: Boolean
+) : AbstractElement(elementName, elementAlias, uniqueIdentifier, attributes, functions, isShell) {
     override val elementKind: ElementKind = ElementKind.ENUM
     override fun getContent(indent: String): String {
         val shellString = if (isShell) DiagramElement.shellString else ""
@@ -48,43 +43,7 @@ $functionsString
 """
     }
 
-    class Builder(override val clazz: KSClassDeclaration, override var isShell: Boolean, override val options: Options, val logger: KSPLogger?) : DiagramElement.Builder<EnumElement> {
-        val companionObject = clazz.declarations.filter { it is KSClassDeclaration && it.isCompanionObject }.map { it as? KSClassDeclaration }.firstOrNull()
-
-        override val extensionProperties: MutableList<KSPropertyDeclaration> = mutableListOf()
-
-        override val allProperties: MutableList<KSPropertyDeclaration>
-            get() = buildList {
-                if (!isShell) {
-                    val validCompanionObjectProperties = companionObject?.getAllProperties()?.filterPropertiesByOptions(clazz, options, logger) ?: emptySequence()
-                    addAll(validCompanionObjectProperties)
-
-                    val validProperties = clazz.getAllProperties().filterPropertiesByOptions(clazz, options, logger)
-                    addAll(validProperties)
-
-                    val otherProperties = clazz.declarations.filterIsInstance<KSPropertyDeclaration>().filterNot { it in clazz.getAllProperties() }.filterPropertiesByOptions(clazz, options, logger)
-                    addAll(otherProperties)
-                }
-
-                addAll(extensionProperties)
-            }.toMutableList()
-
-        override val extensionFunctions: MutableList<KSFunctionDeclaration> = mutableListOf()
-
-        override val allFunctions: MutableList<KSFunctionDeclaration>
-            get() = buildList {
-                if (!isShell) {
-                    val validCompanionObjectFunctions = companionObject?.getAllFunctions()?.filterFunctionsByOptions(clazz, options, logger) ?: emptySequence()
-                    addAll(validCompanionObjectFunctions)
-
-                    val validFunctions = clazz.getAllFunctions().filterFunctionsByOptions(clazz, options, logger)
-                    addAll(validFunctions)
-
-                    val otherFunctions = clazz.declarations.filterIsInstance<KSFunctionDeclaration>().filterNot { it in clazz.getAllFunctions() }.filterFunctionsByOptions(clazz, options, logger)
-                    addAll(otherFunctions)
-                }
-                addAll(extensionFunctions)
-            }.toMutableList()
+    class Builder(clazz: KSClassDeclaration, isShell: Boolean, options: Options, logger: KSPLogger?) : AbstractElementBuilder(clazz, isShell, options, logger) {
 
         override fun build(): EnumElement? {
             return if (options.isValid(clazz, logger)) {
